+++ packages/build-scripts/src/core/Context.ts
@@ -381,8 +381,8 @@ class Context {
         };
       }
       const plugins: [string, IPluginOptions] = Array.isArray(pluginInfo) ? pluginInfo : [pluginInfo, undefined];
-
-      const pluginPath = require.resolve(plugins[0], { paths: [this.rootDir] });
+      const pluginResolveDir = process.env.PLUGIN_DIR ? [process.env.PLUGIN_DIR, this.rootDir] : [this.rootDir];
+      const pluginPath = path.isAbsolute(plugins[0]) ? plugins[0] : require.resolve(plugins[0], { paths: pluginResolveDir });
       const options = plugins[1];
 
       try {